requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
idna==2.10
loguru==0.5.3
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
urllib3==1.25.11
win32-setctime==1.0.3
